add make rpm for fetcherfs
remove temp code
